+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## 2.21 branch
 ### 2.21.0
 * numerous code adjustments for improvement and adaptation to PHP 7.4 new features;
+* `OptionsParser` now uses the `$defaults` array property to store default values;
 * requires at least PHP 7.4.
 
 ## 2.20 branch